+++ magnebot/magnebot.py
@@ -26,11 +26,13 @@ from magnebot.actions.drop import Drop
 from magnebot.actions.reset_arm import ResetArm
 from magnebot.actions.reset_position import ResetPosition
 from magnebot.actions.rotate_camera import RotateCamera
+from magnebot.actions.look_at import LookAt
+from magnebot.actions.move_camera import MoveCamera
 from magnebot.actions.reset_camera import ResetCamera
 from magnebot.actions.slide_torso import SlideTorso
 from magnebot.actions.stop import Stop
 from magnebot.actions.wait import Wait
-from magnebot.constants import TDW_VERSION
+from magnebot.constants import TDW_VERSION, DEFAULT_CAMERA_POSITION_TORSO, DEFAULT_CAMERA_POSITION_COLUMN
 from magnebot.wheel import Wheel
 
 
@@ -173,12 +175,15 @@ class Magnebot(RobotBase):
     _CHECKED_VERSION: bool = False
 
     def __init__(self, robot_id: int = 0, position: Dict[str, float] = None, rotation: Dict[str, float] = None,
-                 image_frequency: ImageFrequency = ImageFrequency.once, check_version: bool = True):
+                 image_frequency: ImageFrequency = ImageFrequency.once, parent_camera_to_torso: bool = True,
+                 visual_camera_mesh: bool = False, check_version: bool = True):
         """
         :param robot_id: The ID of the robot.
         :param position: The position of the robot. If None, defaults to `{"x": 0, "y": 0, "z": 0}`.
         :param rotation: The rotation of the robot in Euler angles (degrees). If None, defaults to `{"x": 0, "y": 0, "z": 0}`.
         :param image_frequency: [The frequency of image capture.](image_frequency.md)
+        :param parent_camera_to_torso: If True, the camera will be parented to the Magnebot's torso. If False, the camera will be parented to the Magnebot's column.
+        :param visual_camera_mesh: If True, the camera will receive a visual mesh. The mesh won't have colliders and won't respond to physics. If False, the camera won't have a visual mesh.
         :param check_version: If True, check whether an update to the Magnebot API or TDW API is available.
         """
 
@@ -233,12 +238,14 @@ class Magnebot(RobotBase):
         """
         self.collision_detection: CollisionDetection = CollisionDetection()
         """:field
-        The current (roll, pitch, yaw) angles of the Magnebot's camera in degrees as a numpy array. This is handled outside of `self.state` because it isn't calculated using output data from the build. See: `Magnebot.CAMERA_RPY_CONSTRAINTS` and `self.rotate_camera()`
+        The current (roll, pitch, yaw) angles of the Magnebot's camera in degrees as a numpy array. This is handled outside of `self.state` because it isn't calculated using output data from the build. See: `magnebot.actions.RotateCamera.CAMERA_RPY_CONSTRAINTS` and `self.rotate_camera()`
         """
         self.camera_rpy: np.array = np.array([0, 0, 0])
         self._previous_resp: List[bytes] = list()
         self._previous_action: Optional[Action] = None
         self._check_version: bool = check_version
+        self._parent_camera_to_torso: bool = parent_camera_to_torso
+        self._visual_camera_mesh: bool = visual_camera_mesh
 
     def get_initialization_commands(self) -> List[dict]:
         """
@@ -461,14 +468,40 @@ class Magnebot(RobotBase):
         # Update the camera RPY angles.
         self.camera_rpy = np.array(self.action.camera_rpy[:])
 
-    def reset_camera(self) -> None:
+    def look_at(self, target: Union[int, Dict[str, float], np.ndarray]) -> None:
         """
-        Reset the rotation of the Magnebot's camera to its default angles.
+        Rotate the Magnebot's camera to look at a target object or position.
+
+        This action is not compatible with `rotate_camera()` because it will ignore (roll, pitch, yaw) constraints; if you use this action, `rotate_camera()` won't work as intended until you call `reset_camera()`.
+
+        :param target: The target. If int: An object ID. If dict: A position as an x, y, z dictionary. If numpy array: A position as an [x, y, z] numpy array.
         """
 
-        self.action = ResetCamera()
+        self.action = LookAt(target=target)
+
+    def move_camera(self, position: Union[Dict[str, float], np.ndarray]) -> None:
+        """
+        Move the Magnebot's camera by an offset position.
+
+        By default, the camera is parented to the torso and will continue to move when the torso moves. You can prevent this by setting `parent_camera_to_torso=False` in the Magnebot constructor.
+
+        :param position: The positional offset that the camera will move by.
+        """
+
+        self.action = MoveCamera(position=position)
+
+    def reset_camera(self, position: bool = True, rotation: bool = True) -> None:
+        """
+        Reset the rotation of the Magnebot's camera to its default angles and/or its default position relative to its parent (by default, its parent is the torso).
+
+        :param position: If True, reset the camera's position.
+        :param rotation: If True, reset the camera' rotation.
+        """
+
+        self.action = ResetCamera(position=position, rotation=rotation, parented_to_torso=self._parent_camera_to_torso)
         # Reset the camera RPY angles.
-        self.camera_rpy = np.array([0, 0, 0])
+        if rotation:
+            self.camera_rpy = np.array([0, 0, 0])
 
     def reset(self, position: Dict[str, float] = None, rotation: Dict[str, float] = None) -> None:
         super().reset(position=position, rotation=rotation)
@@ -477,6 +510,7 @@ class Magnebot(RobotBase):
         self.camera_rpy: np.array = np.array([0, 0, 0])
         self.collision_detection = CollisionDetection()
         self._previous_resp.clear()
+        self.commands.append(self._get_parent_avatar_command())
 
     def slide_torso(self, height: float) -> None:
         """
@@ -512,16 +546,16 @@ class Magnebot(RobotBase):
                               {"$type": "set_pass_masks",
                                "pass_masks": ["_img", "_id", "_depth"],
                                "avatar_id": self.static.avatar_id},
-                              {"$type": "parent_avatar_to_robot",
-                               "position": {"x": 0, "y": 0.053, "z": 0.1838},
-                               "body_part_id": self.static.arm_joints[ArmJoint.torso],
-                               "avatar_id": self.static.avatar_id,
-                               "id": self.static.robot_id},
+                              self._get_parent_avatar_command(),
                               {"$type": "enable_image_sensor",
                                "enable": False,
                                "avatar_id": self.static.avatar_id},
                               {"$type": "set_img_pass_encoding",
                                "value": False}])
+        # Visualize the camera.
+        if self._visual_camera_mesh:
+            self.commands.append({"$type": "add_visual_camera_mesh",
+                                  "avatar_id": self.static.avatar_id})
 
     def _set_dynamic_data(self, resp: List[bytes]) -> None:
         """
@@ -557,3 +591,14 @@ class Magnebot(RobotBase):
                 "position": self.initial_position,
                 "rotation": self.initial_rotation,
                 "id": self.robot_id}
+
+    def _get_parent_avatar_command(self) -> dict:
+        """
+        :return: A command to parent the avatar to the Magenbot. The parent body part is determined by `self._parent_camera_to_torso` and the `parent_camera_to_torso` constructor parameter.
+        """
+
+        return {"$type": "parent_avatar_to_robot",
+                "position": DEFAULT_CAMERA_POSITION_TORSO if self._parent_camera_to_torso else DEFAULT_CAMERA_POSITION_COLUMN,
+                "body_part_id": self.static.arm_joints[ArmJoint.torso if self._parent_camera_to_torso else ArmJoint.column],
+                "avatar_id": self.static.avatar_id,
+                "id": self.static.robot_id}